+++ src/commands/editAccessRequestConfigCommand.ts
@@ -1,0 +1,24 @@
+import { TenantTreeItem } from "../models/IdentityNowTreeItem";
+import * as vscode from 'vscode';
+import { getResourceUri } from "../utils/UriUtils";
+import { openPreview } from "../utils/vsCodeHelpers";
+
+export class EditAccessRequestConfigCommand {
+
+    async execute(node: TenantTreeItem): Promise<void> {
+        console.log("> EditAccessRequestConfigCommand.execute", node);
+
+        const publicIdentitiesConfigUri = getResourceUri(node.tenantName,
+            'access-request-config',
+            null,
+            "Access Request Configuration");
+
+        await vscode.window.withProgress({
+            location: vscode.ProgressLocation.Notification,
+            title: 'Opening Access Request Configuration...',
+            cancellable: false
+        }, async (task, token) => {
+            await openPreview(publicIdentitiesConfigUri)
+        });
+    }
+}