+++ src/extension.ts
@@ -55,6 +55,7 @@ import { FormDefinitionImporterTreeViewCommand } from './commands/form/FormDefin
 import { WorkflowExportCommand } from './commands/workflow/WorkflowExportCommand';
 import { WorkflowImporterTreeViewCommand } from './commands/workflow/WorkflowImporterTreeViewCommand';
 import { EditPublicIdentitiesConfigCommand } from './commands/editPublicIdentitiesConfigCommand';
+import { EditAccessRequestConfigCommand } from './commands/editAccessRequestConfigCommand';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -93,7 +94,13 @@ export function activate(context: vscode.ExtensionContext) {
 		vscode.commands.registerCommand(commands.EDIT_PUBLIC_IDENTITIES_CONFIG,
 			editPublicIdentitiesConfigCommand.execute,
 			editPublicIdentitiesConfigCommand));
-			
+
+	const editAccessRequestConfigCommand = new EditAccessRequestConfigCommand()
+	context.subscriptions.push(
+		vscode.commands.registerCommand(commands.EDIT_ACCESS_REQUEST_CONFIG,
+			editAccessRequestConfigCommand.execute,
+			editAccessRequestConfigCommand));
+
 	const identityNowDataProvider = new IdentityNowDataProvider(context, tenantService);
 	vscode.window.registerTreeDataProvider(commands.TREE_VIEW, identityNowDataProvider);